requirements.txt
# Core Python Requirements
python-dotenv>=0.21.0
requests>=2.28.1
pyyaml>=6.0

# Ollama Integration
ollama>=0.1.0
transformers>=4.30.0

# Twitter Client Dependencies
selenium>=4.10.0
webdriver-manager>=3.8.6

# Optional: Enhanced Functionality
schedule>=1.2.0
beautifulsoup4>=4.12.2

# Logging and Monitoring
loguru>=0.7.0

# Optional: Natural Language Processing
nltk>=3.8.1

# Development and Testing
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0